TET-327 removed hook and moved logic to useEffect
